+++ tests/test_checklog.py
@@ -8,6 +8,7 @@ import unittest
 from click.testing import CliRunner
 
 from acsoo.checklog import checklog
+from acsoo.main import main
 
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
@@ -45,4 +46,17 @@ class TestChecklog(unittest.TestCase):
         ])
         assert res.exit_code == 0
         expected = "errors that did not cause failure (2):"
+        assert expected in res.output
+
+    def test4(self):
+        runner = CliRunner()
+        res = runner.invoke(main, [
+            '-c', os.path.join(DATA_DIR, 'test_checklog.cfg'),
+            'checklog',
+            os.path.join(DATA_DIR, 'test1.log'),
+        ])
+        assert res.exit_code != 0
+        expected = "errors that caused failure (1):"
+        assert expected in res.output
+        expected = "errors that did not cause failure (1):"
         assert expected in res.output